Rename `newConmonOCIRuntime` to `newOCIRuntime`
The so-called "shim-v2" OCI runtimes (notably Kata) do not follow the
usual command-line interface like `runc` or `crun`. Support for the
command-line interface was dropped with Kata v2, and this was
intentional. Patches to restore that functionality were rejected by
the Kata architecture committee, so there is no chance of that coming
back.

The Kata community still see themselves as "OCI runtimes", with the
rationale that a runtime supporting the shim v2 interface does not
need to also support the command-line interface. Going through an
RPC-only interface vastly simplifies the state management in the
runtime, avoiding a complex system of state save/restore across
command-line instances. Instead, a single process acts both as a shim
for the underlying VM's I/Os and as a runtime, taking its commands not
from the command-line but from an RPCV channel.

Whether a runtime can legitimately give up on the command-line
interface is debated, and point of views differ on whether Kata
actually follows the OCI spec. In any case, experience shows that they
won't budge. Furthermore, they work with `containerd` and `cri-o`, and
in terms of command-line support, are perfectly satisfied with
`nerdctl` or `docker`.

Adding support for a Shim-V2 runtime like Kata means that we won't
talk to `conmon` anymore, and that we need to leave a persistent
process behind to respond to gRPC, instead of the usual trick used by
`podman` of passing exit arguments on the command-line that would
cause the runtime to call `podman` back.