Replace "conmon" with "shim" in generic code
With the introduction of a shim-v2 interface, we now have two distinct
shim interfaces, but they have a lot in common, e.g. a pid, a path,
and so on.

Replace references to conmon that refer to the shim and not really
specifically to conmon with "shim" instead of "conmon". For example,
replace `ConmonInfo` with `ShimInfo`.